% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysquares.R
\name{mysquares}
\alias{mysquares}
\title{Squaring function}
\usage{
mysquares(x)
}
\arguments{
\item{x}{A vector of numbers}
}
\value{
A vector of numbers squared
}
\description{
Squaring function
}
\examples{
a <- seq(0,100,10)
mysquares(a)
}
